31/1/2024

+++ nodejs/src/controller/price_controller.js
@@ -1,5 +1,6 @@
 import { validationResult } from "express-validator";
 import priceService from "../service/price_service";
+import bedTypeService from "../service/bedType_service";
 
 const getPriceByBedType = async (req, res) => {
     try {
@@ -49,14 +50,18 @@ const insertPrice = async (req, res) => {
 }
 
 const updatePrice = async (req, res) => {
+    const validate = validationResult(req);
+    if (!validate.isEmpty()) {
+        return res.status(400).json({ error_code: validate.errors[0].msg })
+    }
     let id, name, hour, day, week, month;
     try {
         id = req.body.id;
-        name = req.body.name==""?null:req.body.name;
-        hour = req.body.hour?null:req.body.hour;
-        day = req.body.day?null:req.body.day;
-        week = req.body.week?null:req.body.week;
-        month = req.body.month?null:req.body.month;
+        name = req.body.name;
+        hour = req.body.hour;
+        day = req.body.day;
+        week = req.body.week;
+        month = req.body.month;
     } catch (error) {
         return res.status(500).json({ error_code: error })
     }
@@ -79,12 +84,18 @@ const updatePrice = async (req, res) => {
 const deletePrice = async (req, res) => {
     try {
         const id = req.body.id;
-        const rs = await priceService.deletePrice(id);
-        if (rs.status) {
-            return res.status(200).json({ result: newprice.result });
-        } else {
-            return res.status(500).json({ error_code: msg })
+        const bedtype=await bedTypeService.findBedTypeByDefaultPrice(id);
+        if(bedtype){
+            return res.status(400).json({ error_code: "Không thể xoá đơn giá mặc định" });
+        }else{
+            const rs = await priceService.deletePrice(id);
+            if (rs.status) {
+                return res.status(200).json({ result: rs.result });
+            } else {
+                return res.status(500).json({ error_code: rs.msg })
+            }
         }
+        
     } catch (error) {
         return res.status(500).json({ error_code: error })
     }
